New flow with TouchID

+++ TradeItIosTicketSDK/TradeItTicketViewSDK/EditViewController.m
@@ -41,7 +41,7 @@
     
     NSArray * actions = [self getOrderActionValues];
     int i;
-    for(i = actions.count - 1; i > 0; i--) {
+    for(i = (int) actions.count - 1; i > 0; i--) {
         if([actions[i] isEqualToString: [[[self tradeSession] orderInfo] action]]) {
             break;
         }
@@ -49,7 +49,7 @@
     [orderTypePicker selectRow:i inComponent:0 animated:NO];
     
     NSArray * types = [self getOrderTypeValues];
-    for(i = types.count - 1; i > 0; i--) {
+    for(i = (int) types.count - 1; i > 0; i--) {
         if([types[i] isEqualToString:[[self tradeSession] orderType]]) {
             break;
         }
